+++ gcovr/tests/coexisting_object_directories-from_build_dir-without_search_dir/source/fn_08.cpp
@@ -1,0 +1,4 @@
+#include <iostream>
+void fn_08() {
+   std::cout << "08" << std::endl;
+}