+++ gcovr/tests/coexisting_object_directories-from_root_dir-without_object_dir/source/fn_02.cpp
@@ -1,0 +1,4 @@
+#include <iostream>
+void fn_02() {
+   std::cout << "02" << std::endl;
+}